+++ src/Adapters/AdapterLoader.js
@@ -17,6 +17,7 @@ export function loadAdapter(adapter, defaultAdapter, options) {
       }
     }
   } else if (typeof adapter === "string") {
+    /* eslint-disable */
     adapter = require(adapter);
     // If it's define as a module, get the default
     if (adapter.default) {
