+++ src/Adapters/Auth/instagram.js
